requirements.txt
# Sign Language Communication Web App - Requirements
# Python 3.8+

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Computer Vision
opencv-python==4.8.1.78

# Machine Learning
tensorflow==2.15.0
keras==2.15.0

# Hand Detection (Optional but Recommended)
mediapipe==0.10.5

# Data Processing
numpy==1.24.3
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0

# Optional: For deployment
gunicorn==21.2.0

# Optional: For better debugging
flask-cors==4.0.0
